Move HyperbolicDiffeo to main code
